build: Upgrade ESLint to v9 and migrate config
