+++ lib/offline/storage.js
@@ -1503,7 +1503,6 @@ shaka.offline.Storage = class {
     try {
       goog.asserts.assert(
           this.networkingEngine_, 'Need networking engine.');
-      // eslint-disable-next-line require-atomic-updates
       mimeType = await shaka.net.NetworkingUtils.getMimeType(uri,
           this.networkingEngine_,
           this.config_.streaming.retryParameters);
