+++ lib/player.js
@@ -6292,7 +6292,6 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
     try {
       goog.asserts.assert(
           this.networkingEngine_, 'Need networking engine.');
-      // eslint-disable-next-line require-atomic-updates
       mimeType = await shaka.net.NetworkingUtils.getMimeType(uri,
           this.networkingEngine_,
           this.config_.streaming.retryParameters);
@@ -8352,7 +8351,7 @@ shaka.Player.TYPICAL_BUFFERING_THRESHOLD_ = 0.5;
  * @define {string} A version number taken from git at compile time.
  * @export
  */
-// eslint-disable-next-line no-useless-concat, max-len
+// eslint-disable-next-line no-useless-concat
 shaka.Player.version = 'v4.12.0' + '-uncompiled';  // x-release-please-version
 
 // Initialize the deprecation system using the version string we just set
