+++ lib/polyfill/video_play_promise.js
@@ -27,7 +27,6 @@ shaka.polyfill.VideoPlayPromise = class {
       const originalPlay = HTMLMediaElement.prototype.play;
       // eslint-disable-next-line no-restricted-syntax
       HTMLMediaElement.prototype.play = function() {
-        // eslint-disable-next-line no-restricted-syntax
         const p = originalPlay.apply(this);
         if (p) {
           // This browser is returning a Promise from play().
