+++ lib/util/config_utils.js
@@ -125,7 +125,7 @@ shaka.util.ConfigUtils = class {
     let last = configObject;
     let searchIndex = 0;
     let nameStart = 0;
-    while (true) {  // eslint-disable-line no-constant-condition
+    while (true) {
       const idx = fieldName.indexOf('.', searchIndex);
       if (idx < 0) {
         break;
