+++ lib/util/platform.js
@@ -795,7 +795,6 @@ shaka.util.Platform = class {
         maxResolution.height = 1080;
       }
     } else if (Platform.isHisense()) {
-      // eslint-disable-next-line new-cap
       if (window.Hisense_Get4KSupportState &&
           // eslint-disable-next-line new-cap
           window.Hisense_Get4KSupportState()) {
