+++ lib/util/string_utils.js
@@ -344,7 +344,6 @@ shaka.util.StringUtils.fromCharCodeImpl_ = new shaka.util.Lazy(() => {
       // This can't use the spread operator, or it blows up on Xbox One.
       // So we use apply() instead, which is normally not allowed.
       // See issue #2186 for more details.
-      // eslint-disable-next-line no-restricted-syntax
       const foo = String.fromCharCode.apply(null, buffer);
       goog.asserts.assert(foo, 'Should get value');
       return foo.length > 0; // Actually use "foo", so it's not compiled out.
@@ -367,7 +366,6 @@ shaka.util.StringUtils.fromCharCodeImpl_ = new shaka.util.Lazy(() => {
           // This can't use the spread operator, or it blows up on Xbox One.
           // So we use apply() instead, which is normally not allowed.
           // See issue #2186 for more details.
-          // eslint-disable-next-line no-restricted-syntax
           ret += String.fromCharCode.apply(null, subArray);  // Issue #2186
         }
         return ret;
