+++ lib/util/tXml.js
@@ -152,7 +152,6 @@ shaka.util.TXml = class {
                 indexOfCloseTag = closeTag.indexOf(shortTag);
               }
             }
-            // eslint-disable-next-line no-restricted-syntax
             if (indexOfCloseTag == -1) {
               const parsedText = S.substring(0, pos).split('\n');
               throw new Error(
