+++ package.json
@@ -27,11 +27,11 @@
     "cspell": "^8.17.1",
     "dialog-polyfill": "^0.5.6",
     "es6-promise-polyfill": "^1.2.0",
-    "eslint": "^8.57.1",
-    "eslint-config-google": "^0.14.0",
-    "eslint-plugin-shaka-rules": "file:./build/eslint-plugin-shaka-rules",
+    "eslint": "^9.18.0",
+    "eslint-config-google": "github:google/eslint-config-google#3ae571a",
     "esprima": "^4.0.1",
     "fontfaceonload": "^1.0.2",
+    "globals": "^15.14.0",
     "google-closure-compiler-java": "^20240317.0.0",
     "google-closure-deps": "^20230802.0.0",
     "google-closure-library": "^20221102.0.0",
