+++ test/mss/mss_player_integration.js
@@ -22,13 +22,8 @@ describe('MSS Player', () => {
   /** @type {!shaka.test.Waiter} */
   let waiter;
 
-  // eslint-disable-next-line max-len
   const url = 'https://playready.directtaps.net/smoothstreaming/SSWSS720H264/SuperSpeedway_720.ism/Manifest';
-
-  // eslint-disable-next-line max-len
   const playreadyUrl = 'https://playready.directtaps.net/smoothstreaming/SSWSS720H264/SuperSpeedway_720.ism/Manifest';
-
-  // eslint-disable-next-line max-len
   const playreadyLicenseUrl = 'https://test.playready.microsoft.com/service/rightsmanager.asmx?cfg=(persist:false,sl:150)';
 
   function checkPlayReadySupport() {
