+++ test/test/boot.js
@@ -55,7 +55,6 @@ function failTestsOnNamespacedElementOrAttributeNames() {
           arg.includes(':')) {
         fail('Use namespace-aware ' + name);
       }
-      // eslint-disable-next-line no-restricted-syntax
       return real.apply(this, arguments);
     };
   };
